ddbb_manager: persist: support for lists, merge instead of add
After the last commit, this should make it possible to deal with several
Python instances representing the same semantic object without duplicates
in the database.

Still, I should implement something to get existing entities from the DDBB
instead of re-creating them (which requires extra calls to the Web3
provider, and thus, is slower).

Signed-off-by: Jose Diaz de Grenu de Pedro <jose.diaz@bosonit.com>